feat(examples-4): Refactor types and properties in Layout Builder
Revised the types and properties of the Layout Builder components. This includes updating function parameters to be of type MouseEvent, replacing property 'layout' with 'flexDirection', and refining 'grow' and 'shrink' types to be CSSProperties. Constructed tags to handle style keys are also introduced. The updating routine is modified to use proper CSS values and needless padding check function is eliminated.