Replace 'value' with 'formula' in SignalState
The 'value' field in the SignalState interface and related areas has been replaced with 'formula'. This change affects both the definition of SignalState and its usage in functions like onAddSignal(), onEditSignal(), and the initialization of state signals. It also simplifies the Notifiable components in SignalDetailDialogPanel. On the same note, irrelevant comments and code have been removed.