Update display property in ToolBar.tsx
The code for the display property in the ToolBar.tsx component has been adjusted. Previously, the display was set to 'flex' by default, but now it is initialized as 'none'. The display setting is then modified to 'flex' when a corresponding element is found, ensuring appropriate visibility of the toolbar.